# openapi/gateway.v1.yaml
openapi: 3.0.3
info:
  title: ApptRemind Gateway API
  version: v1
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/business/services:
    post:
      summary: Create service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessServiceCreateRequest"
            examples:
              consult:
                value:
                  name: "Consultation"
                  duration_minutes: 30
                  price: 25
                  description: "Initial consult"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
              examples:
                created:
                  value:
                    id: "c6b6b7e0-7c2a-4a07-8a9f-1b2c3d4e5f60"
    get:
      summary: List services
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessService"
              examples:
                list:
                  value:
                    - id: "c6b6b7e0-7c2a-4a07-8a9f-1b2c3d4e5f60"
                      business_id: "9f5f9e1a-7f8d-4b9c-9f7b-1e8f0c1d2e3f"
                      name: "Consultation"
                      duration_minutes: 30
                      price: "25.00"
                      description: "Initial consult"
                      created_at: "2026-01-28T10:00:00Z"
...
  /api/v1/appointments:
    get:
      summary: List appointments
      security:
        - bearerAuth: []
      parameters:
        - name: business_id
          in: query
          required: false
          schema:
            type: string
          description: Optional override; otherwise derived from JWT.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentSummary"
              examples:
                list:
                  value:
                    - appointment_id: "b285af7d-8710-4549-b762-2a28d25ca690"
                      staff_id: "2d7f53f6-0b5f-4d0d-8f49-7a9c3b3b9c2a"
                      service_id: "c6b6b7e0-7c2a-4a07-8a9f-1b2c3d4e5f60"
                      start_time: "2026-01-28T14:00:00Z"
                      end_time: "2026-01-28T14:30:00Z"
                      status: "booked"
                      created_at: "2026-01-28T10:10:00Z"
...
  /api/v1/billing/checkout/session:
    get:
      summary: Get checkout session status (public)
      parameters:
        - in: query
          name: session_id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingCheckoutSessionStatus"
              examples:
                completed:
                  value:
                    session_id: "cs_test_123"
                    tier: "starter"
                    status: "completed"
                    updated_at: "2026-01-28T10:10:00Z"
                    completed_at: "2026-01-28T10:10:00Z"
  /api/v1/billing/checkout/session/ack:
    post:
      summary: Acknowledge checkout return (public, token-protected)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingCheckoutAckRequest"
            examples:
              ack:
                value:
                  session_id: "cs_test_123"
                  state: "return-token"
                  result: "success"
      responses:
        "200":
          description: OK
...
  /api/v1/billing/webhooks/stripe:
    post:
      summary: Stripe webhook (signature verified)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BillingCheckoutSessionStatus:
      type: object
      properties:
        session_id:
          type: string
        tier:
          type: string
        status:
          type: string
        updated_at:
          type: string
        completed_at:
          type: string
        canceled_at:
          type: string
        expired_at:
          type: string
    BillingCheckoutAckRequest:
      type: object
      required: [session_id, state, result]
      properties:
        session_id:
          type: string
        state:
          type: string
          description: return token embedded into success/cancel URL
        result:
          type: string
          description: success | cancel
...
    BusinessServiceCreateRequest:
      type: object
      required: [name, duration_minutes, price]
      properties:
        name:
          type: string
        duration_minutes:
          type: integer
        price:
          type: number
          format: float
        description:
          type: string
    BusinessService:
      type: object
      properties:
        id:
          type: string
        business_id:
          type: string
        name:
          type: string
        duration_minutes:
          type: integer
        price:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
...
    IdResponse:
      type: object
      properties:
        id:
          type: string
...
    AppointmentSummary:
      type: object
      properties:
        appointment_id:
          type: string
        staff_id:
          type: string
        service_id:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        status:
          type: string
        cancelled_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time